Rename dashboard app.
