+++ .babelrc
@@ -13,7 +13,6 @@
     ["transform-class-properties", { "spec": true }],
     ["module-resolver/lib/index.js", {
       "alias": {
-        "react": "preact",
         "": "./src"
       }
     }],
