+++ docs/apps/dashboard/user/index.html
@@ -40,23 +40,23 @@
                 </a>
               
                 
-                <a class="source" href="../account.html">
-                  apps/account/account.js
+                <a class="source" href="../dashboard.html">
+                  apps/dashboard/dashboard.js
                 </a>
               
                 
                 <a class="source" href="../index.html">
-                  apps/account/index.js
+                  apps/dashboard/index.js
                 </a>
               
                 
                 <a class="source" href="index.html">
-                  apps/account/user/index.js
+                  apps/dashboard/user/index.js
                 </a>
               
                 
                 <a class="source" href="../users/index.html">
-                  apps/account/users/index.js
+                  apps/dashboard/users/index.js
                 </a>
               
                 
@@ -72,6 +72,11 @@
                 
                 <a class="source" href="../../main/login/index.html">
                   apps/main/login/index.js
+                </a>
+              
+                
+                <a class="source" href="../../main/login/login.html">
+                  apps/main/login/login.js
                 </a>
