+++ docs/elements/not-found.html
@@ -40,23 +40,23 @@
                 </a>
               
                 
-                <a class="source" href="../apps/account/account.html">
-                  apps/account/account.js
+                <a class="source" href="../apps/dashboard/dashboard.html">
+                  apps/dashboard/dashboard.js
                 </a>
               
                 
-                <a class="source" href="../apps/account/index.html">
-                  apps/account/index.js
+                <a class="source" href="../apps/dashboard/index.html">
+                  apps/dashboard/index.js
                 </a>
               
                 
-                <a class="source" href="../apps/account/user/index.html">
-                  apps/account/user/index.js
+                <a class="source" href="../apps/dashboard/user/index.html">
+                  apps/dashboard/user/index.js
                 </a>
               
                 
-                <a class="source" href="../apps/account/users/index.html">
-                  apps/account/users/index.js
+                <a class="source" href="../apps/dashboard/users/index.html">
+                  apps/dashboard/users/index.js
                 </a>
               
                 
@@ -72,6 +72,11 @@
                 
                 <a class="source" href="../apps/main/login/index.html">
                   apps/main/login/index.js
+                </a>
+              
+                
+                <a class="source" href="../apps/main/login/login.html">
+                  apps/main/login/login.js
                 </a>
