+++ package.json
@@ -45,6 +45,7 @@
   "dependencies": {
     "@app-elements/form": "1.3.8",
     "@app-elements/helmet": "2.0.1",
+    "@app-elements/list-resource": "^2.0.19",
     "@app-elements/loading-indicator": "1.0.4",
     "@app-elements/modal": "2.0.0",
     "@app-elements/notification": "1.0.4",
