+++ src/apps/dashboard/dashboard.js
@@ -3,20 +3,20 @@ import Router, { RouteTo, Link } from '@app-elements/router'
 // import { getState } from '/store'
 import { routes } from './index'
 
-const AccountHeader = () =>
+const DashboardHeader = () =>
   <header className='alt'>
     <Link name='users'>Users</Link>&nbsp;
     <Link name='login'>Login</Link>
   </header>
 
-export default function AccountApp () {
+export default function DashboardApp () {
   // const { token } = getState()
   const token = 1
   return token == null
     ? <RouteTo name='login' />
     : (
-      <div id='account-layout'>
-        <AccountHeader />
+      <div id='dashboard-layout'>
+        <DashboardHeader />
         <Router routes={routes} />
       </div>
     )
